Ensure that the the generated airflow.cfg contains a random jwt_secret and fernet_key (#46966)
* Ensure that the the generated airflow.cfg contains a random jwt_secret and fernet_key

I don't know exactly when this got broken, and unforutnaltey it wasn't tested,
but I suspect it might have been around the time we swapped the default config
from a config file to the yaml based values. I.e. a while ago!

To make sure it doesn't get broken I've gone and added some unit tests

And to make my next PR and test easier I have done the same thing with the
`auth_jwt_secret` that we do for fernet_key -- of only set it in the config
file if we're generating that file, not always in memory.

* Improve upgrade path by generating and warning about the missing config