Update documentation on how to setup loca venv and PyCharm IDE (#47017)
This is not yet a final setup, as we will be introducing more
distributions and packages, but this one describes the "current"
state of how to create your local venv, how to keep it up-to-date
(both with `uv` - simpler, and `pip` - more complex) and how to
work with providers. Also it shows how to configure PyCharm/IntelliJ
in order to effectively debug and test stuff in local venv.

Again - that will still change a bit - for example `task_sdk` will
also need `tests` to be added to your `test sources root` and
`airflow` core will move to separate sub-folder but i think it's
good to update the docs to reflect current state and to review the
docs to remove all old references.