Ensure that the Execution API's '/{ti}/run' end point is idempotent. (#47126)
If there are any network issues between the client and the API server, it was
possible that the API could receive the request, update the state to RUNNING,
but the client never gets the response. When this happens the client would
retry the HTTP request, but the retry would fail with a HTTP 409 Conflict
error as the state was already marked as running.

This fixes that by accepting the request again, as long as it is from the same
host+unixname+pid we saw already.

Co-authored-by: Mehul Goyal <mehul.goyal@astronomer.io>