fix: don't use blocking property access for async purposes (#47326)
The AwsGenericHook provides the `@property` decorated function `async_conn`
for building and accessing the async boto3 client. Unfortunately, this
is results in blocking calls to the Airflow Db or secrets backends within async contexts.

This PR provides an async method `get_async_conn` as an alternative that can be awaited
in async contexts. This method will call the underlying sync code in a
`sync_to_async` wrapper.

The `async_conn` property is now deprecated and will be removed in future versions.