Handle var/conn not-found cases better when defined at top-level (#47863)
closes: https://github.com/apache/airflow/issues/47862


When a var/conn is defined at top level in a dag and it does't exist, the error is pretty trash. Fixing that to handle that case better.

Used the same as in the linked issue, but error is better now:
```
Traceback (most recent call last):
  File "/opt/airflow/task-sdk/src/airflow/sdk/definitions/variable.py", line 50, in get
    return _get_variable(key, deserialize_json=deserialize_json).value
  File "/opt/airflow/task-sdk/src/airflow/sdk/execution_time/context.py", line 108, in _get_variable
    raise AirflowRuntimeError(msg)
airflow.sdk.exceptions.AirflowRuntimeError: VARIABLE_NOT_FOUND: {'key': 'hi'}
```