Replace DAGs with dags in docs (#47959)
Reading "DAGs" everywhere in the docs is ugly, unnatural, and unnecessary. We don't always need to make a, perhaps a bit pedantic, reference to the class name when talking about dags. And, as discussed on the dev list, there seemed to be a general agreement that reference to the "directed acyclic graph" is not particularly helpful. So, here, I change most instances of DAGs in the docs to dags.