Remove change_sensor_mode_to_reschedule from base executor (#48649)
There are no longer any single process executors so this field is
not needed.

This allows us to remove a runtime dependency of the task runner
on loading the executor class when running sensors which used to
override the prepare_for_execution to load the executor class to
check this field which is False in all supported executors.