Fix local running of backcompat tests after we switched to UV (#48669)
We were including a real `airflow/__init__.py` which has the unfortunate
side-effect of making `AIRFLOW_V_3_0_PLUS` find 3.0.0dev0 as the version!
Oops. The fix for that is to mount an empyt init file instead (we need a file
to keep the uv build etc happy)

Also uninstall edge provider for 2.9, as that needs 2.10.

Oh and volume mount devel-common as that is part of test code.