Allow bundle_name and relative_fileloc to be None API response (#48779)
This is necessary because, in practice, they are null in the db. And, it does not make much sense to correct this at upgrade time because, they're not necessarily even defined in dags anymore, and we don't know which bundle they'd be a part of post upgrade. So, let's just leave it null, and it will be resolved by dag processor or through airflow dags reserialize.